Implements user defined module support in Node.js enabling require('foo') for a module in the user's project
This is accomplished by switching to using projection buffers for .js files inside of our project. At put in a hidden header at the top of the file
Which includes a <reference path=”…”> tag. This is a reference to a file we generate in the temp dir which includes our API definition for
Node.js as well as the code from the user's project. When the file changes on disk we'll go off and regenerate the reference file for the
project which causes the JS engine to make the new completions available.

To generate the reference file we take the file's code and surround it in a function like:

function user_mod() {
exports = {};

// USER CODE GOES HERE such as:

function foo() {
}

exports.foo = foo;

// END USER CODE

return exports;
}

This causes the JS analysis engine to properly see the newly assigned values to the exports table, and we get completion against
the user module.

Adds test cases for intellisense against Node.js built-ins, a user defined module, as well as a few test cases for the projection buffer support.

Also changes the default project so that it includes the web application defaults so that we're configured to use a custom web server. This prevents the web application project from attempting to reconfigure IIS express when opening the project which can be a slow process.