#3 Confusing error message when expanding an empty array: Check for zero-length arrays and provide a dummy No children child item instead of VS-generated Children could not be evaluated.
#41 Problems when renaming a file in project:
Caused by picking up an imported <Content> item generated by our .targets from the actual <Compile> item in the .njsproj file. The imported item is associated with the hierarchy node, and its Include attribute is what gets changed when renaming. Since it is imported, it is never actually persisted on save. The fix is to ignore imported items for the purpose of generating hierarchy nodes.

#54 Drag & drop folder doesn't work:
Caused by a bug in the code processing child items of a node. It enumerates all siblings, and for each one calls itself recursively, also asking that recursive call to enumerate all siblings. For 3+ nodes, this breaks, because we end up going over the same node more than once.

#61 Children not sorted in debugger view:
Sort the array of children before returning it to debugger.