Fix http://nodejstools.codeplex.com/workitem/117:
These changes add special processing of breakpoints with locations fixed up by the node engine. Specifically, On breakpoint hit for node fixed up breakpoints, we attempt to rebind a breakpoint, once . If we successfully rebind without fixup, we keep the rebind and silently continue. This effectively works around bug 117. This takes advantage of an undocumented feature of the node engine.

It is worth noting that these changes do not include telling vs about fixup in the case that rebind results in fixup, again. For these scenarios we break on the fixup line without having the breakpoint indicator in the editor gutter moving to a fixed up line. This is something we can fix post alpha.

Related unit test added.