Improve performance of adding an existing directory hierarchy
The core of this change is using a dictionary to lookup filenames by full path, and maintaining that dictionary for files and folders. But this includes a lot of cleanups to MPFProj to get this right. A lot of the cleanups are focused around getting the right split between ProjectNode and HierarchyNode as there was some event handling which wasn't clear if it made sense or not. The end result is that HierarchyNode loses a ton of VS interfaces which were never called on HierarchyNode, they were only ever called on a ProjectNode.

One of the methods it loses is the Close method which now becomes a void returning method instead of having COM legacy in its signature.

FindChild splits into several different methods depending upon the usage:
HierarchyNode.FindImmediateChildByName – this finds a child node based upon the component of the filename we're looking for.
HierarchyNode.FindImmediateChild – this takes a predicate to find the child based upon whatever attributes you want. Several manual walks of a node's children comparing w/ the caption were updated to use this (which is more correct, because caption need not match the path component name)
ProjectNode.FindNodeByFullPath – this is the replacement that most callers were wanting to do, this returns an on-disk node by filename

Previously Hierarchy nodes all had a hierarchyEventSinks collection on them, and when we'd go to deliver an event (OnItemDeleted, OnItemAdded, Redraw, etc…) we would end up going to one event sink or another. I would expect it to always have been the project nodes events; if we ever were delivering them to a random HierarchyNode there were certainly no subscribers. So all of this event stuff moved from HierarchyNode to ProjectNode. There's also more state that was on HierarchyNode but really only needed to be on the ProjectNode – that includes the parentHierarchy and item ID.

I also killed off VirtualNodeName entirely. It was mainly being used to track the name of folders and it was doing so in an odd way, so now we just track folder names in Url like we would for a file node. And then we have a new interface, IDiskBasedNode, which we use to get the URL of folders or files. Elsewhere where it's used it's not doing anything interesting.

I also killed off docCookie from HierarchyNode. Turns out it's only ever used on FileNode's, so other nodes shouldn't pay the cost.

I also called off lastChild – this just wasn't being used.

I've also ported the existing Python project tests over to Node.js. This also exposed a few spots where we were calling Visual Studio with the wrong project node (not the aggregated project node).

Another perf boost for this is that I made CreateFileNode stop searching for duplicate items. Callers are supposed to figure this out themselves if they need to.

Finally, it adds a fast path for when we're doing the common case of adding an existing directory hierarchy.