Fix for https://nodejstools.codeplex.com/workitem/258
Removes hack to avoid surfacing node rethrow of unhandled exceptions. Note that Node no longer does a rethrow in versions 9.5 (30-Dec-2012) and later. For those who are using earlier versions, rethrows will be surfaced with a callstack but no source. In Beta we will mitigate this by sucking down source for such callstacks.