Fix https://nodejstools.codeplex.com/workitem/223
Change browser launcher to no longer timeout and launch the browser after a minute. Note that there is a short circuit mechanism which will cause the browser launcher to evaporate if the debug session (or node process for start without debugging) goes away before the launcher detects the port opened and launches the browser.