Fix interaction between breakpoint location fixup, breakpoint predicates (hit count, condition) and entry point. This fixes the following bugs and related scenarios:
 - https://nodejstools.codeplex.com/workitem/397
 - conditional breakpoints don't work
 - https://nodejstools.codeplex.com/workitem/330
 - Changing condition of a conditional breakpoint does not cause it to hit

It is worth noting that it does not fix:

 - https://nodejstools.codeplex.com/workitem/395
 - Breakpoints are moved away from valid locations

The basic mechanism is to rebind any breakpoint which had location fixup once, when hit, and only surface the break (hit event) if the rebind binds to the current line where broken. If it does not rebind to the current line, we auto resume to continue stepping or running. There is special work to delay setting predicates on fixed up breakpoints until rebind so the predicates don't filter out the hit event we use to trigger rebind. The changes also include not surfacing breakpoint bound events to VS until rebind for fixed up breakpoints. This keeps breakpoints fixed up for lambda-eval from jumping twice. It does not result in hollow glyphs for breakpoints until fully bound. I am investigating getting hollow glyphs working.

I am continuing to work on https://nodejstools.codeplex.com/workitem/395, looking from empirical clues from Node (V8) that it will need more than one rebind to get a breakpoint binding to steady state.

I have more test I am working on which are not part of this code review.