Tactical fix for https://nodejstools.codeplex.com/workitem/411
Our recent changes to use a node.js specific language guid and content type rather than javascript's broke this scenario. The issue is that when we expose our downloaded scripts to the debugger UI as IDebugDocument*, the debugger UI ignores our declared language service and simply uses the document's moniker's extension to derive language and the Javascript language is used. We refuse requests to create breakpoints other than in node.js files. The tactical fix is to check if a breakpoint request's associated IDebugDocument* is ours. If it is, we allow a breakpoint request, even if the language guid is not ours. These changes get breakpoints in downloaded code working for alpha. Longer term I would like to try making downloaded scripts hidden elements in our project. This should allow full node.js language service binding and also has the promise of getting breakpoints in downloaded scripts persisting across debug sessions.

The changes also include work to get not yet bound breakpoints showing up in the UI (editor gutter) as hollow.