Alpha bug fixes:
https://nodejstools.codeplex.com/workitem/342
https://nodejstools.codeplex.com/workitem/396
https://nodejstools.codeplex.com/workitem/374

Fixes for 342 and 396 improve discoverability and documentation of remote debugging.

The fix(es) for 374 cover three issues I uncovered last night:

1) Breakpoints in builtin open new editor each time they are hit

This was a subtle regressions of debugger code refactoring which broke our caching of documents per loaded module.

2) Breakpoints in node_modules don't bind

This was further fallout of moving away from handling the javascript language guid. The fix for alpha is to allow setting breakpoints in files tagged with the javascript language guid. Post alpha we will do work to further clean this up when we get breakpoints in downloaded to persist across debug sessions.

3) Opening downloaded node_modules from callstack yields odd source

This is an artifact of how we do fuzzy filename matching when attach (vs. launch). The problem is that we match only by whole path or leafname. This causes issues where we arbitrarily match the last of multiple matching leafnames. The fix is to favor partial matches with more matching characters at the end of a partially matching path.