#674 Support remote debugging of Azure web sites
Add a websocket HTTP request handler that wraps and proxies a TCP connection to the assembly deployed to Azure with the web sites.

Add a web.config transform for the Debug configuration that registers and exposes said request handler.

Add support for web socket endpoints to remote debugger - connection strings are now URIs (tcp:// for raw connections, ws:// and wss:// for WebSocket with or without SSL) rather than just plain hostname:port, though the old format is still supported.

Add Attach to debugger (node.js) command to Azure web site nodes in Server Explorer that provides a convenient shortcut for the above functionality.