#748 Attach Debugger command not visible in Server Explorer with VS Update 2
The contracts assembly for Azure websites VS API has changed in SDK 2.3 (which is also in VS Update 2) in a way that breaks existing clients, forcing us to support one of the other. This change updates our code so that it uses the new interfaces. The contract assembly in question is going to be stable from 2.3 RC onwards. Reflection is used to access the interfaces to avoid taking a compile-time dependency on private builds from the 2.3 branch with updated interfaces.

This effectively disables the Attach debugger (node.js) command in Server Explorer for earlier SDK versions (2.3 CTP 2 and below). The remote attach functionality will still be available there via Process -> Attach to Server, by directly entering a ws:// URL.