#778 Breakpoints are not hit when remote debugging
When doing fuzzy filename matching, if there is more than one candidate at the point where local and remote filesystems begin to diverge, pick the one nearest to the point of divergence, rather than failing to match.

Change fuzzy matching test so that it has to deal with both local and remote filenames being duplicated.