#846 Debugger crashes
This is another sweeping fix for async void methods leaking expected exceptions - they need to handle everything that can occur during normal operation or forced detach from DebuggerClient.SendRequestAsync, which includes OperationCanceledException as well as IOException. The try/catch is now refactored to a separate helper method on DebuggerClient, since this pattern is used a lot.

Bound breakpoint removal is made synchronous, as we don't want to report success if we couldn't actually remove it.

Project build for TypeScript E&C is now synchronous on save. This avoids leaking any build exceptions on a threadpool thread, and means that user can't race the source push anymore. The code now also checks whether the build actually succeeded before trying to push the updated .js to debuggee, so a failed build will now be reported as failure to update in the status bar.