#1113 REPL should not give completions when line is a command
Disable completions if current text buffer is a REPL command that starts with ., and add a UI test for that.

Add a basic UI sanity test for REPL completions.

Separate require() completion in its own method to improve readability.

Fix NullReferenceException in NodejsProjectNode.Dispose, if the node is disposed before it is fully initialized (e.g. because initialization fails).