#1012 Comment/uncomment command not available
Suport comment block and uncomment block in Node.js. Most of this code is (almost) identical to the Python Tools counterpart as commenting isn't fundamentally different.

A testcase was also added to Python Tools and the expected/actual results on assertions were backwards in the tests causing logging to be incorrect.