#823 Project from existing code does not build
Cause of the issue: When the project file is in a separate directory of the source and caused the paths to be incorrect. This change ensures that the paths are absolute. I have also added a test to make sure we try to build a project like this and verify that it has no build errors.