#1161 Watch dir for MAX_PATH & warn users
Add a check for file and directory names in project directory or project output directory that are too long, and trigger that check for package management operations (npm UI manager, REPL .npm command, and auto-install from package.json on package creation). If the check fails, provide links that describe various aspects of the problem, and offer to run npm dedup on the project.