#1233 Cannot set as startup project a file that was just renamed .js
This fixes an issue where name changing to a .js does not give the ability to set the file as the startup file. This is because rename was not causing us to re-evaluate the node type. We now ask for a new item node to make sure we convert to the correct ItemNode type during rename.