#1365 Breakpoints inside functions don't work when attaching
Fix fuzzy (filesystem-walking) filename matcher that is used for remote debugging such that it doesn't produce false matches for modules that are not part of the project (such as built-in modules), causing misplacement of breakpoints.

Update existing unit test for the matcher that is supposed to test this case so that it actually does test it properly.