#1216 Cancel saves the selection on Tools->options->Text Editor->node.js->Formatting
This change addresses the issue that when you click cancel on tools options dialogs the checkboxes that you checked stay checked for the next time, despite not being submitted.

This change make certain that every time we open the page, we sync the state of the UI with the settings stored in the backing fields.