#1315 No way to use semicolon in environment variable
Replace <EnvironmentVariables> project property (with semicolon-separated elements) to <Environment> property (with newline-separated elements). The new name is consistent with VC++ projects.

Add project upgrade support that replaces <EnvironmentVariables> with <Environment>.

Use multiline textbox to edit environment variables in the project properties page, and move it to the end of the dialog for better visual arrangement.

Fix test failures due to run without debugging being broken for tests, because PressAnyKey.exe was not being copied to TestResults\Out.