#1171 recognize arguments  Adds basic support for arguments. Adds new ArgumentsValue object which represents  the arguments object. It supports indexing to get the argument values out of the  object. When we analyze a function it goes to the current functions argument set for  the analysis.
#1235 requirejs member info missing
 requirejs was doing an assignment to a var in the outer scope from an inner scope.
 This wasn't updating the correct variable. Fixes scope assignment so we assign to
 the correct scope.

Also stops returning the union for strings which are added together and instead returns
an empty string. Returning the union results in completions which shouldn't actually
be there as we didn't really do the add. Returning an empty string preserves the correct
string completions but prevents the incorrect completions from being added.