#1403 npm ui fails to parse package list
The npm UI is currently broken for everyone who tries to refresh the catalog at the moment.

The issue is that there is a new entry in the package catalog that uses an array instead of a string to represent the homepage, so we fail to parse it. So far it only looks like it's the readline-prompter package that's causing the error, and I haven't seen any entries with multiple homepages, but this fix should take care of it just in case it does happen by representing homepages as lists instead, and listing out all homepages in the package details view.

We also now catch ArgumentExceptions, and log the error and package name to the output window to prevent a similar issue from breaking the entire UI experience in the future - we do not attempt to display the problematic package.