#1402 Azure Worker Role fails to publish with error in ServiceDefinition.csdef
The port variable is not necessary and fails if we don't have an endpoint. Since worker roles don't start with endpoints this fails by default.

Also, in the migration to the wizard, we didn't add the execution of the startup script.

Finally, while editing worker role templates I modified the template to not serve a web file as without an endpoint it can't do anyway.