#1346 Inserting a semicolon in an incorrect place causes an exception to activity log
The FixStatementIndentation method was throwing an ArgumentOutOfRangeException as IndexOfAny throws this if you hand it a negative count to look at. We should check the count before IndexOfAny is called.

Also, it turns out the indentation for ';' characters in invalid sequences had an error where in indented scopes we added unnecessary whitespace.