#1410 TypeScript build options not respected
Typescript build options are not respected due to a bug in TypeScript. The implementation for adding the property groups in TypeScript is to first find the target and put them in a property group of a condition above it. Since we didn't have the target it kept appending them to the end of the file after our target and they weren't getting used for build. By adding a dummy property the properties will be added at the appropriate place.