#1283 Show jsdoc comments in signature tooltips
Enables special handling for /** */ comments in the scanner, and uses it in expression parsing to propagate to assignments, function declarations and expressions, and variable initializers, ultimately exposed as documentation on function signatures.

Add @param parsing for signatures, extracting documentation strings and information about optionality and default values, for display in signature help.