#1533 Allow $(ProjectDir) in paths in NTVS properties window
We were changing the UI to show the evaluated properties which would wipe out the unevaluated options on save. We were also previously not allowing a user to save if we thought they had an error. I removed that check since the user should be allowed to save if they want. We warn them at least.