#1594 Error reading package.json
Previously we loaded the package.json file with shared read priveleges. This change makes sure we have exclusive access which makes sure the file is in a steady state. Logic was also added to retry around IOExceptions and other exceptions that can occur while trying to open the file when unavailable.