#1632 npm - Searching for grunt-contrib-uglify doesn't return the module
The problem was that '-' is used as a NOT operator in FTS.

The fix is to replace hyphens with spaces. It's somewhat hacky, but the only negative effect is that we can't use the NOT operator in searches, which isn't a common (or even documented) scenario to begin with. Other options included was moving away from FTS, which would a) hurt perf, b) be risky, c) ruin search for other scenarios like separated word queries. Another option was using phrase search by default, but a) couldn't get it to work even though it seems possible according to a stackoverflow answer, and b) that would prevent us from using regex altogether