Fixes a bug in the ImportSettings where we didn't properly create the njsproj file.
Adds a new command to the REPL, .project. This cleans up the syntax of the .npm. No more [project]. It also allows future REPL commands to target specific projects. Per-project behavior is up to the command to decide if it will participate.
- .project <name of project>
The command matches the name to those of loaded projects. If there is no matching node.js project it informs the user of that. If a match is found its set
- .project
This will display the current setting.