#1740 Exception when formatting document
The underlying issue to this failure is that our tokencache.map was not being updated correctly on inserts. We would reset the cache for the current line and count - 1 following lines in the case of the line split. By resetting map[index+count] (when that exists as a map entry) we avoid this issue.

We check the size because sometimes this final line doesn't exist in the _map to delete. It isn't reliable, but holding ctrl-v while pasting the short repro for this bug causes us to try to remove a non-existent row from the array without it. It isn't common that a user would edit that fast, but just in case we should guard against it.