Fix exception in classifier
This appears to be the root cause of issue #130.  The use of `<=` allows
for `currentLineTemp` to be equal to the number of lines in the
snapshot.  When that occurs and the value is passed to `TokenizeLine` it
will result in an exception because it will be passed to
`GetLineFromLineNumber` which throws when the line requested is `>=` to
the line count.

closes #130