Use generated, strongly typed resource file in Nodejs
**Defect**
To get string resources, we currently, access everything through an SR class. This class hardcodes a bunch of strings to lookup resources in a resx by name. Hardcoding these strings is not great because these strings can get out of sync with the actual resx file.

**Fix**
This hooks up the resx file to generate an typed cs file automatically. This ensures that the resx file and the string resource uses cannot get out of sync.