Improve performance of repl window output
Currently, printing large structures to the repl window can take a very long time. I believe the primary cause is we end up appending output almost one character at a time, repeating the same operations over and over again.

This change updates ReplWindow to take a set of outputs to be appended instead of a single output. This allows moving some of the computations out of the inner loops and reusing previously computed values.