Limit Execution time of IsMissing check for #679
Investigating #679. I can't repo the problem, but from the provided dump, believe that one work around may be to limit how long Directory.Exists spends executing. This call may hang for network drives in certain cases.

This logic adds a 250ms timeout to the call. If the call times out or the files does not exists, we return that the package is missing.