Include typings d.ts files in Dev14 Projects Automatically
There are two problems with Intellisense in Dev14 Update 2, using typescript for Javascript Intellisense:
* You have to rename the d.ts files to trigger Intellisense to work. Just including the file in the project is not enough
* typings files are not automatically picked up or added to the project.

This change fixes both of these issues.

The first for the first is triggering an `OnItemAdded` to get an included d.ts file noticed by type script

The second part is fixed by auto including any file node that is in the typings directory.