Always install node.d.ts typing file
This makes sure we install node.d.ts for intellisense, even if it is not explicitly required by another installed package.

We should merge this change in because it makes things better, but there is currently an issue where `OnIdleNodeModules` is never fired for projects without any node modules. I'm looking into handling that, but the proper fix may involve changing the existing behavior of the code.

closes #770