Allows NTVS repl window to work for any project type
We currently only support running `.npm` commands in node projects for the Repl window. This change allows us to run repl window commands in other project types as well.

Since these projects may not have a `package.json` file for npm to use, also added a command to create it.