Fix duplicate TypeScriptCompile entries being created
**Bug**
In NTVS 1.2 Alpha, we end up calling IncludeInProject much more often for automatic type acquisition. There seems to be a bug in typescript where this ends up modifying the solution to include a new TypeScriptCompile entry every time the file is reincluded. This results in the project file being changed every time the project is loaded. I do not know if there are any perf problems or bugs introduced by having duplicate TypeScriptCompile entries.

**Fix**
Only run `IncludeInProject` if the file is not already included.