Limit Depth of created dependecy nodes
This PR further addresses memory usage.

**Bug**
Right now, we end up parsing the entire node dependecy tree and creating objects to represent every dependency in it. This uses a lot of memory and a lot of time during boot.

The use of this is that we allow users to explore the entire module dependency tree in VS directly.

**Fix**
These operations are negatively impacting user experiance, especially when loading projects with lots of dependencies (such as the eslint source). I also am not convinced that seeing the dependencies of your project dependenices is actually useful, at least 99.99% of the the time.

This fix limits the depth to which we create dependency nodes to 1 (only root deps). Memory usage and boot time are much improved.