Fix problems when sending debug commands after Node program runs to completion
**Bug**
We currently pause on the last line of a program. Node 6 made some change and now does not respond to debug messages sent after this pause.

**Fix**
This change has two fixes:

* Add an explicit timeout when removing breakpoints. We have an identical timeout when adding breakpoints so it probably makes sense to add this.
* On our low level send calls, register the cancellation token to cancel the promise.

A better fix would be just to never send requests to node after we run to the last line, but it is not clear how we can detect this without a wrapper. The debug protocol does not seem to emit any events about this and the app is offically still running.

closes #935