Fix smart indent settings crash
Issue #1014

**Bug**
Crash in `SmartIndent.cs` in the interactive window. I believe that the root cause is #995, which changed how we store our settings.

**Fix**
Switch to using the NodeJs language preferences from typescript instead. This ensures we are using the same logic to get the language settings as we use in other places.

closes #1014