Fix Node 6 Debugging
Issue #934

**Bug**
Initial breakpoints currently sometimes do not resolve when using Node 6.x. In most cases, you can remove the breakpoint and add it again later to get it to resolve.

**Fix**
I believe something with Node 6 changed the casing for the names of scripts in the v8-debug protocol.

There were two fixes here:

* Use a case insensitive regular expression to match the script names when setting a breakpoint (I believe this is similar to what VS code does)
* Fix one case where we were relying on a case sensitive comparision to update a breakpoint.

This is not quite tested to be get merged in just yet, but wanted to see if there were any initial concerns or feedback with the approch here.