Add Typings to Express 4 Templates auto install from pre-existing typings.json
Issue #1037

**Bug**
This fix addresses the two `serve-static` related errors we are seeing int #1037. The root cause is that the express typings file reference other typings files that are not installed by default. For a template, this results in two errors when creating a new project and a pretty bad first experiance.

**Fix**
Two parts

* Update typings acquisition to automatically attempt to install typings that are already listed in a `typings.json` file, if such a file exists.
* Add a `typings.json` file to the Express4 template.

I'll add similar `typings.json` files to the other express 4 based JavaScript templates if this is the approch we decide we want to go with.