Get msbuild path from VSTarget version in BuildRelease.ps1
Bug
The BuildRelease.ps1 script can only be run from a visual studio command prompt. It needs a specific version of msbuild, and the default one found by a normal powershell / normal command prompt will not work.

Fix
Lookup the msbuild path based on the vstarget version we are using for the build. This checks the registry to find where `msbuild.exe` is located.

closes #1101