Prototype Starter Express App template Express 3 to 4 change
Related to #1058

This prototypes moving the basic StarterApp from Express3 to Express4. The update is based on our existing Express 4 template.

I'd like to get feedback on this change before doing the work to migrate the other 5 templates. If we can cut down the number of templates, that would also help us a lot.