Extract functions in Diagnostics command
We're looking to add some additional project info to our Diagnostics / Telemetry (#1116). I've took a quick pass through the current diagnostics window to clean up the code a bit and hopefully make this easier.

The cleanup extracted functionality for logging each of the major info groups, such as project or assemblies. This just organizes and cleans up the code so that adding new fields should be easier.