Port Node Debug provider to NTVS
This change takes the existing node debugger provider embedded in the Open Folder team's code base, and moves it into the NTVS code. This will allow us to maintain the debug provider going forward.

Testing
Tested on VS15. There seems to be a bug with anycode where `ProviderPriority` is not respected, so the builtin copy of the node debug provider is always used over this new debug provider in the NTVS codebase. I changed the file extension type to test this change, and saw this debug provider being used.