Remove GlobalModulesNode
**Bug**
We currently populate a npm node with a list of global modules for every NTVS project. For users with a few globals installed and/or lots of NTVS projects in a solution, this can lead to perf issues with both memory and CPU, since the global modules node includes the entire dependency tree of modules.

**Fix**
Simply remove this node.

Closes #1125