Delete Starter Express App Templates
**Bug**
The starter express app templates are very out of date and do not provide a good example of best practices while developing a modern node web application. We do not have the resources to keep them up to date. We also do not want to ship templates that are out of date and promote bad design (for example, we still are referencing Jade in all of these templates, which has been renamed to Pug).

If we ever do want to provide a more feature-rich template for new users, starting from scatch would be better IMO would be better than trying to update all these templates.

**Fix**
Just delete all these templates. Any projects created from these templates will continue to work fine, this only effects the new project experiance.

We would want to take this into NTVS 1.3 and probably VS15 Preview 5.