Fix Building NTVS with devenv /Build (#1236)
* Fix Building NTVS with devenv /Build

Issue #1230

**Bug**
Crash when trying to build an NTVS project from the command line with `devenv /build`. The root cause is in the `UIThread` where we attempt to call `TaskScheduler.FromCurrentSynchronizationContext`. In non-ui flows, this throws an `InvalidOperationException`.

**Fix**
If `TaskScheduler.FromCurrentSynchronizationContext` fails, try setting the current syntronization context manually and retrying the call. This fixes the `UIThread`, but also fixes all other code paths (including some in the VS extension bases themseleves) that use `TaskScheduler.FromCurrentSynchronizationContext` or assume a SyncronizationContext exists.

closes #1230

* Better scope fix