Remove Dockerfile Template (#1237)
* Remove Dockerfile Template

**Bug**
I believe at some point, NTVS planned to integrate with Docker. This does not appear to have really happened, but we still have this weird DockerFile template hanging around. This file template is not a general dockerfile, but NTVS specific. I do not believe we should include it anymore.

**Fix**
Delete.

* Remove docker file from setup files list