Copy Dlls for Test Adapter into VSIX for VS15
**Bug**
Test adapter is not working in VS15.

One cause is that it cannot resolve two dlls: shell.interop and newtonsoft.json. The first is in the GAC in vs14 and the second is installed by our MSI (not entirely sure how it works for devbuilds)

**Fix**
Copy these dlls into the vsix explicitly. I tried embedding shell.interop, but it cannot be embedded.