Delete Add Website Templates (#1234)
* Delete Add Website Templates

**Bug**
In the best case, we see under 10% of users creating new node website projects from `new Website` flow vs the `new project` flow (10% number taken for `NodeJsWebapp` vs `AddWebsiteNodejsWebapp`. Other templates compare less favorably, with a grand total of 1 users over the past 10 days creating a AzureExpressApp through the new website flow). This suggests to me that this flow is not discoverable. For NTVS, `new website` also adds zero actual value since the templates are exactly the same as with `new project`.

The cost is that we have another copy of templates to maintain and another set of flows we have to cover in our test passes. And we only present an incomplete subset of node website templates in this flow to begin with.

**Fix**
Just delete these templates.

* Update csproj