Remove Node.js Editors
**Bug**
We are moving to fully use Typescript to power our editors. #1255 removed the option to switch to the old ES5 IntelliSense in the UI, but users can still reach this by switching to the old Node.js editor. We do not want this behavior.

**Fix**
Delete the old Node.js edtiors and editor factories entirely.