Limit Depth of Npm Modules
This is a resbumission of #944

**Bug**
Now that our Node analyzer cannot be used in the main editor, there no real reason why we need to build a complete tree of npm installs. As #944 documented, this is a major source of memory and cpu usage at boot

**Fix**
Limit depth of explored npm tree to a single level. This brings some very nice perf gains for projects that have complex depdency tress.