Remove old Node.js Editor and Instead Use Typescript Editor (#1260)
* Remove ES5 IntelliSense Option

As explained in #1254, there are few reasons for anyone to switch back to the old ES5 based intellisense. Our telemtry also shows that very very few people are switching back too.

This change removes the option to switch from the VS entirely. This locks all main editors in ES6/Typescript backed mode. We'll keep around the old analysis engine code a bit longer in order to power the Interactive window, but will eventually rip it out too.

Closes #1254

* Remove unused intellisense level changed logging call

* Remove two more unused options and disable intellisense page on vs15

* Remove horizontal divider from options page

* Remove Node.js Editors

**Bug**
We are moving to fully use Typescript to power our editors. #1255 removed the option to switch to the old ES5 IntelliSense in the UI, but users can still reach this by switching to the old Node.js editor. We do not want this behavior.

**Fix**
Delete the old Node.js edtiors and editor factories entirely.

* Small cleanup for diagnosticscommand

* Remove unused TelemtryEvent name

* Remove Update contentType function