Remove Global Registry NodeJs InstallDir fallback
**Bug**
With VS15 (and prior versions too), we do not want to write data to a global registry location. We currently persist and `InstallDir` field to the NTVS registry.

**Fix**
Remove this fallback. This code is used to lookup files from the NTVS package. We should always use the files relative to the node.js tools dll instead.