Remove old node analyzer and intellisense (#1280)
* Remove Old Node Analysis and Intellisense/Analysis Code

**bug**
With NTVS 1.3, we have removed the old Node.js editor. This means that the old analyzer and intellisense logic are no longer required. This code has been the source of a lot of problems and we do want to maintain our own implementation of this. We instead rely on TypeScript for this functionality. TypeScript provides

**fix**
Delete all the analysis logic and anything and everything connected with this.

This change does remove some functionality. Perhaps most significantly, IntelliSense and syntax highlighting are no longer supported in the repl window (both work just fine in the editor). I beleive that the positives of this change for the entire NTVS project outweigh these regressions.

* Fix build scripts and few other references to analyzer

* Revert ruleset

* REvert assembly version

* Fix a few more references to jsanalysistest and remove unsued vars

* Remove some unused ui tests

* Remove two more analysis.dll references

* Remove reference generator

* Remove a few more references to reference generator

* Fix reference test

* Remove some unused constants

* Remove analysis level enum

* Trigger typing acq correctly on install