Provide ability to add project references
This change updates NodejsProjectNode to detect if the UWP project
flavor is active to provide support for project references.
Alternative approach could have been to make NodejsProjectNode public
(so that I could reference it from the UWP project system)
but that would cause a huge changes as opposed to this less
intrusive change.