Pass VSInstall Root to Test Adapter
**Bug**
In VS15, unit tests are currently not being discovered. It looks like one problem is that when we load the njsproject file, MSBuildExtensionsPath32 is set to incorrectly  look in the global location, instead of the new location inside of VS.

**fix**
Pass VSInstallRoot from the main nodejs instance to the test discoverer. Use this environment variable to manually set MSBuildExtensionsPath32